Extract router to a separate file
